organization: move AddImageButton to its own file
